@@ -6,10 +6,10 @@
     "homepage": "http://developers.google.com/api-client-library/php",
     "license": "Apache-2.0",
     "require": {
-        "php": ">=5.6"
+        "php": "^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7||^8.5.13"
+        "phpunit/phpunit": "^9.6"
     },
     "autoload": {
         "psr-4": {
